Create playable version
In here I have created a version we can play with a constant that controls board size, it creates a list that it can then check back on to see if the user hit a bomb or not. function board_two needs work as it right now is just laying out the numbers with nothing else so this is probably the main thing to do. It also needs a couple of more tweaks like directions at the start and a way for it to know that the game is over